feat: admin full control — create/edit events for organisers, edit user profiles
- Extract organiser event wizard into shared components/event/EventFormWizard
  so the admin portal reuses the same create/edit flow
- Admin API: POST /api/admin/events (create on behalf of organiser),
  GET+PATCH /api/admin/events/[id] (edit any status; live events stay live),
  GET+PUT /api/admin/users/[id] (profile edit + Cognito email sync)
- Allow admin sessions in /api/upload for event media
- Admin UI: /admin/events/create + /admin/events/[id]/edit pages, Create/Edit
  buttons on the events list, EditUserDialog on the users page
- Dedicated AdminNavBar for the admin portal
- Audit log actions CREATE_EVENT / EDIT_EVENT / EDIT_USER; middleware now
  protects all admin paths
- E2E coverage in e2e/admin-full-control.spec.ts

Closes #74
