+++ app/admin/audit/page.tsx
@@ -19,6 +19,8 @@ interface AuditLog {
 const ACTION_COLORS: Record<string, string> = {
   APPROVE_EVENT:       "bg-primary/10 text-primary",
   REJECT_EVENT:        "bg-red-400/10 text-red-400",
+  CREATE_EVENT:        "bg-primary/10 text-primary",
+  EDIT_EVENT:          "bg-blue-400/10 text-blue-300",
   DELETE_EVENT:        "bg-red-400/10 text-red-400",
   PIN_EVENT:           "bg-blue-400/10 text-blue-300",
   UNPIN_EVENT:         "bg-white/[0.05] text-muted",
@@ -31,6 +33,7 @@ const ACTION_COLORS: Record<string, string> = {
   ACTIVATE_ORGANISER:  "bg-primary/10 text-primary",
   BAN_USER:            "bg-red-400/10 text-red-400",
   UNBAN_USER:          "bg-primary/10 text-primary",
+  EDIT_USER:           "bg-blue-400/10 text-blue-300",
   REFUND_REGISTRATION: "bg-amber-400/10 text-amber-300",
 };
 
@@ -54,14 +57,15 @@ function metaSummary(meta: Record<string, unknown> | null): string {
   if (typeof meta.athleteName === "string") parts.push(meta.athleteName);
   if (typeof meta.amountCents === "number") parts.push(`$${((meta.amountCents as number) / 100).toFixed(2)}`);
   if (typeof meta.newStatus   === "string") parts.push(`→ ${meta.newStatus}`);
+  if (Array.isArray(meta.fields) && meta.fields.length) parts.push(`${meta.fields.join(", ")} changed`);
   return parts.join(" · ");
 }
 
 const ALL_ACTIONS = [
-  "APPROVE_EVENT", "REJECT_EVENT", "DELETE_EVENT", "PIN_EVENT", "UNPIN_EVENT",
+  "APPROVE_EVENT", "REJECT_EVENT", "CREATE_EVENT", "EDIT_EVENT", "DELETE_EVENT", "PIN_EVENT", "UNPIN_EVENT",
   "BULK_APPROVE", "BULK_REJECT", "BULK_DELETE",
   "VERIFY_ORGANISER", "UNVERIFY_ORGANISER", "SUSPEND_ORGANISER", "ACTIVATE_ORGANISER",
-  "BAN_USER", "UNBAN_USER",
+  "BAN_USER", "UNBAN_USER", "EDIT_USER",
   "REFUND_REGISTRATION",
 ];
