+++ app/admin/layout.tsx
@@ -1,4 +1,4 @@
-import NavBar from "@/components/NavBar";
+import AdminNavBar from "@/components/admin/AdminNavBar";
 import AmplifyProvider from "@/app/organiser/AmplifyProvider";
 import { AuthProvider } from "@/context/AuthContext";
 
@@ -6,7 +6,7 @@ export default function AdminLayout({ children }: { children: React.ReactNode })
   return (
     <AmplifyProvider>
       <AuthProvider>
-        <NavBar />
+        <AdminNavBar />
         {children}
       </AuthProvider>
     </AmplifyProvider>
